voice_management_system/requirements.txt
Django==1.11.4
djangorestframework==3.6.3
django-rest-auth==0.9.1
factory-boy==2.9.2
django-cors-headers==2.1.0
